requirements.txt
tweepy
configparser
pandas
python-dotenv
nltk
pyspark
findspark

